==> 02_eda_open_rate.py <==
# -*- coding: utf-8 -*-
"""02_eda_open_rate

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1EzLRg-I5lVSAqEa4cbYqQohOj0ZgHCwJ
"""



"""## 1. Email open rate"""

delivered_count_df = user_event_df[['user_uuid','group_name']+PO_number_list]
for e in PO_number_list:
    delivered_count_df[e] = delivered_count_df[e].notnull()

delivered_count_df = delivered_count_df.groupby('group_name').sum().reset_index()
delivered_count_df.head()

open_count_df = user_event_df[['user_uuid','group_name']+PO_number_list]
for e in PO_number_list:
    open_count_df[e] = open_count_df[e] == 'open'

open_count_df = open_count_df.groupby('group_name').sum().reset_index()
open_count_df.head()

open_rate_df = open_count_df
for e in PO_number_list:
    open_rate_df[e] = open_count_df[e]/delivered_count_df[e]

open_rate_df.head(20)

result = open_rate_df.round(2).set_index('group_name')
fig, ax = plt.subplots(figsize=(20, 16))
clean = result.apply(pd.to_numeric, errors="coerce")   # non-parsable values → NaN
sns.heatmap(clean, annot=True, fmt=".2f", cmap="viridis", ax=ax)


plt.xticks(fontsize=20)
plt.yticks(fontsize=20)
plt.title('Heatmap of Email opening rate',fontsize=22)


plt.show()

#neg_count_df = user_event_df.drop(columns=['user_uuid'])
neg_count_df = user_event_df[['user_uuid','group_name']+PO_number_list]
for e in PO_number_list:
    neg_count_df[e] = ((neg_count_df[e] == 'spamreport') | (neg_count_df[e] == 'unsubscribe'))

neg_count_df = neg_count_df.groupby('group_name').sum().reset_index()
neg_count_df.head()
