==> 05_ab_test.py <==
# -*- coding: utf-8 -*-
"""05_ab_test

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1uGiC666SWTbMbzVaNBl8ID48gaBMXuO4
"""



exp_funding_merge_df['p_val'] = exp_funding_merge_df.apply(cal_pval,axis=1)
exp_funding_merge_df['significant'] = exp_funding_merge_df['p_val']<=0.05
exp_funding_merge_df.head(50)

temp =exp_funding_merge_df.sort_values('delta_fund_rate',ascending=False)
temp[["funding_rate", "funding_rate_in_control",'delta_fund_rate']] = temp[["funding_rate", "funding_rate_in_control",'delta_fund_rate']]*100
ax = temp.plot(x="group_name", y=["funding_rate", "funding_rate_in_control"], kind="bar",figsize=(16, 12))

plt.xticks(fontsize=20)
plt.yticks(fontsize=20)
plt.ylabel('funding rate (%)',fontsize=20)
plt.title('funding rate (sorted by delta)',fontsize=22)

plt.legend(fontsize=20)
plt.grid(which="both", color="0.65", linestyle="-")

ax2=ax.twinx()
ax2.plot(temp['group_name'], temp['delta_fund_rate'],color="green",marker="o")

ax2.set_ylabel('size')
plt.yticks(fontsize=20)
plt.ylabel('delta funding rate (%)',fontsize=20)

plt.show()

temp =exp_funding_merge_df[exp_funding_merge_df['link_rate']<1].sort_values('delta_link_rate',ascending=False)
temp[["link_rate", "link_rate_in_control",'delta_link_rate']] = temp[["link_rate", "link_rate_in_control",'delta_link_rate']]*100

ax = temp.plot(x="group_name", y=["link_rate", "link_rate_in_control"], kind="bar",figsize=(16, 12))

plt.xticks(fontsize=20)
plt.yticks(fontsize=20)
plt.ylabel('link rate (%)',fontsize=20)
plt.title('Link rate (sorted by delta)',fontsize=22)

plt.legend(fontsize=20)
plt.grid(which='both', color='0.65', linestyle='-')

ax2=ax.twinx()
ax2.plot(temp['group_name'], temp['delta_link_rate'],color="green",marker="o")

#ax = temp['delta_fund_rate'].plot(secondary_y=True, color='k', marker='o',figsize=(16, 12))
ax2.set_ylabel('size')
plt.yticks(fontsize=20)
plt.ylabel('delta link rate (%)',fontsize=20)

plt.show()
